# alzawaj-project-backend/src/config/openapi.yaml
openapi: 3.0.0
info:
  title: Zawag Islamic Marriage Platform API
  version: 1.0.0
  description: API documentation for the Zawag Islamic Marriage Platform
servers:
  - url: http://localhost:5001
    description: Development server
  - url: https://alzawaj-backend.onrender.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MarriageRequest:
      type: object
      properties:
        id:
          type: string
          example: "6051d2d7e1b8a20015d23f5c"
        senderId:
          type: string
          example: "6051d2d7e1b8a20015d23f5a"
        receiverId:
          type: string
          example: "6051d2d7e1b8a20015d23f5d"
        status:
          type: string
          enum: [pending, accepted, rejected]
          example: "pending"
        message:
          type: string
          example: "Assalamu alaikum, I am interested in getting to know you better inshallah."
        createdAt:
          type: string
          format: date-time
          example: "2021-03-17T10:30:00Z"
        expiresAt:
          type: string
          format: date-time
          example: "2021-04-17T10:30:00Z"
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "بيانات التسجيل غير صحيحة"
        statusCode:
          type: integer
          example: 400
        error:
          type: array
          items:
            type: string
          example:
            - "تأكيد كلمة المرور مطلوب"
            - "الاسم يجب أن يكون بين 2-100 حرف"
            - "العمر يجب أن يكون بين 18-100 سنة"
            - "حالة اللحية مطلوبة للذكور"
            - "الحالة المالية مطلوبة للذكور"
            - "نوع السكن مطلوب للذكور"
security:
  - bearerAuth: []
...
paths:
...
  /api/requests/sent:
    get:
      tags: [Requests]
      summary: Get sent marriage requests
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MarriageRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/requests/cancel/{requestId}:
    post:
      tags: [Requests]
      summary: Cancel a sent marriage request
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Marriage request successfully cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/admin/stats:
    get:
      tags: [Admin]
      summary: Get admin stats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/admin/users/action:
    post:
      tags: [Admin]
      summary: User actions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                action:
                  type: string
                  enum: [suspend, activate, delete, verify]
                reason:
                  type: string
      responses:
        '200':
          description: User action successfully processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...